+++ Mile.Helpers/Mile.Helpers.CppWinRT.cpp
@@ -10,6 +10,8 @@
 
 #include "Mile.Helpers.CppWinRT.h"
 
+#include <Windows.h>
+
 #include <winrt/Windows.ApplicationModel.h>
 #include <winrt/Windows.ApplicationModel.Resources.Core.h>
 #include <winrt/Windows.Foundation.Collections.h>
@@ -68,4 +70,32 @@ winrt::hstring Mile::WinRT::GetLocalizedString(
     winrt::hstring const& ResourcePath)
 {
     return Mile::WinRT::GetLocalizedString(ResourcePath, ResourcePath);
+}
+
+winrt::hresult_error Mile::WinRT::ToHResultError()
+{
+    try
+    {
+        throw;
+    }
+    catch (winrt::hresult_error const& ex)
+    {
+        return ex;
+    }
+    catch (std::bad_alloc const&)
+    {
+        return winrt::hresult_error(E_OUTOFMEMORY);
+    }
+    catch (std::out_of_range const& ex)
+    {
+        return winrt::hresult_out_of_bounds(winrt::to_hstring(ex.what()));
+    }
+    catch (std::invalid_argument const& ex)
+    {
+        return winrt::hresult_invalid_argument(winrt::to_hstring(ex.what()));
+    }
+    catch (std::exception const& ex)
+    {
+        return winrt::hresult_error(E_FAIL, winrt::to_hstring(ex.what()));
+    }
 }